Added cards onto explore page
