+++ components/environment-card.js
@@ -5,7 +5,7 @@
  *                                  
  *
  * #HOW TO CALL:
- *      <SumaryCard     data />
+ *      <EnvironmentCard     data />
  *
  *    @prop { Array }   data  - object data with the statitics information
  *
@@ -55,6 +55,12 @@ const useStyles = makeStyles(theme => ({
         alignItems: 'center',
 
     },
+    posCard: {
+        // position: 'absolute',
+        // right: '1vw',
+        // top: '7vh',
+        // zIndex: 1,
+    }
 }));
 
 export default function EnvironmentCard(props) {
@@ -81,9 +87,6 @@ export default function EnvironmentCard(props) {
                         <div className={classes.labelData}>
                             <Image src="/CO2.svg" alt="CO2" width={30} height={30}/>
                         </div>
-                        {/* <Typography variant="h6" component="div" className={classes.numData}>
-                            {annual_potential} kW/h
-                        </Typography> */}
                         <Typography sx={{ fontSize: 12 }} className={classes.labelData} gutterBottom>
                         {annual_potential} Metric Tons / kWh
                         </Typography>
@@ -92,9 +95,6 @@ export default function EnvironmentCard(props) {
                         <div className={classes.labelData}>
                             <Image src="/Seedling.svg" alt="Seedling" width={30} height={30}/>
                         </div>
-                        {/* <Typography variant="h6" component="div" className={classes.numData}>
-                            {annual_potential} kW/h
-                        </Typography> */}
                         <Typography sx={{ fontSize: 12, textAlign: 'center' }} className={classes.labelData} gutterBottom>
                         {annual_potential} Metric Tons / Urban trees planted
                         </Typography>
@@ -103,94 +103,17 @@ export default function EnvironmentCard(props) {
                         <div className={classes.labelData}>
                             <Image src="/Car.svg" alt="Car" width={30} height={30}/>
                         </div>
-                        {/* <Typography variant="h6" component="div" className={classes.numData}>
-                            {annual_potential} kW/h
-                        </Typography> */}
                         <Typography sx={{ fontSize: 12}} className={classes.labelData} gutterBottom>
                         {annual_potential} Metric / Per Mile
                         </Typography>
                     </div>
                 </ Box>
-                {/* <Box
-                    sx={{
-                        p: 0.5,
-                        m: 0.5,
-                    }}
-                    className={classes.dataRow}
-                >
-                    <div className={classes.groupData}>
-                        <Typography variant="h6" component="div" className={classes.numData}>
-                            {annual_potential} kW/h
-                        </Typography>
-                        <Typography sx={{ fontSize: 12, textAlign: 'center'}} className={classes.labelData} gutterBottom>
-                        Metric Tons / Urban trees{<br />}planted
-                        </Typography>
-                    </div>
-                    <div className={classes.groupData}>
-                        <Typography variant="h6" component="div" className={classes.numData}>
-                            {annual_potential} kW/h
-                        </Typography>
-                        <Typography sx={{ fontSize: 12 }} className={classes.labelData} gutterBottom>
-                        Metric / Per Mile
-                        </Typography>
-                    </div>
-                </ Box> */}
-                {/* <Box
-                    sx={{
-                        p: 0.5,
-                        m: 0.5,
-                    }}
-                    className={classes.dataRow}
-                >
-                    <div className={classes.groupData}>
-                        <Typography variant="h6" component="div" className={classes.numData}>
-                            {hourly_potential} kW/h
-                        </Typography>
-                        <Typography sx={{ fontSize: 12 }} className={classes.labelData} gutterBottom>
-                            Hourly Potential
-                        </Typography>
-                    </div>
-                    <div className={classes.groupData}>
-                        <Typography variant="h6" component="div" className={classes.numData}>
-                            {daily_potential} kW/h
-                        </Typography>
-                        <Typography sx={{ fontSize: 12 }} className={classes.labelData} gutterBottom>
-                        Daily Potential
-                        </Typography>
-                    </div>
-                </Box>
-                <Box
-                    sx={{
-                        p: 0.5,
-                        m: 0.5,
-                    }}
-                    className={classes.dataRow}
-                >
-                    <div className={classes.groupData}>
-                        <Typography variant="h6" component="div" className={classes.numData}>
-                            {avail_rooftop} m<sup style={{
-                                                        fontSize:'small'
-                                                    }}>2</sup>
-                        </Typography>
-                        <Typography sx={{ fontSize: 12, textAlign: 'center' }} gutterBottom>
-                            Available roof area {<br />} for installation
-                        </Typography>
-                    </div>
-                    <div className={classes.groupData}>
-                        <Typography variant="h6" component="div" className={classes.numData}>
-                            {num_panels}
-                        </Typography>
-                        <Typography sx={{ fontSize: 12 }} className={classes.labelData} gutterBottom>
-                        Panels for Installation
-                        </Typography>
-                    </div>
-                </Box> */}
             </CardContent>
         </React.Fragment>
     );
 
     return (
-        <Box sx={{ minWidth: '35%' }}>
+        <Box sx={{ minWidth: '35%' }} className={classes.posCard}>
             <Card >{card}</Card>
         </Box>
     );
